4개 상품 크롤링데이터 추가

+++ src/main/java/com/imchobo/sayren_back/domain/crawling/service/CrawlingService.java
@@ -141,7 +141,7 @@ public class CrawlingService {
         product = productRepository.save(Product.builder()
                 .name(name)
                 .description(description)
-                .price(price)
+                .price((long)price)
                 .productCategory(category)
                 .modelName(modelName)
                 .isUse(false)
